Some refactor and body parts nbt data not found handling
