+++ src/main/java/xyz/srgnis/bodyhealthsystem/mixin/OnPlayerJoinMixin.java
@@ -10,7 +10,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import xyz.srgnis.bodyhealthsystem.BHSMain;
-import xyz.srgnis.bodyhealthsystem.body.PlayerBodyProvider;
+import xyz.srgnis.bodyhealthsystem.body.player.PlayerBodyProvider;
 import xyz.srgnis.bodyhealthsystem.body.player.PlayerBody;
 
 @Mixin(PlayerEntity.class)
